Day 19: Only use resources and robots in `states` map
We'll still use the `minutes_elapsed` field as the value so we can
determine when to skip a state. We won't skip a state if the current
time is earlier than the one we examined earlier (because it might be
that we can make some moves that will improve our outcome). But if we've
already spent as much time or more, there's no way that we can do better
so let's just skip this state.